[#2] feat: 메인페이지에 검색바 추가 및 기존 검색바 shared로 리팩토링
