+++ src/features/player-search/index.ts
@@ -1,4 +1,4 @@
-export { SearchBar } from './ui/SearchBar'
+export { SearchBar } from '../../shared/ui/SearchBar'
 export { FilterPanel } from './ui/FilterPanel'
 export { ViewToggle } from './ui/ViewToggle'
 export { Pagination } from './ui/Pagination'
